Amelioration

+++ src/main/java/org/psychohelp/psychohelp/service/ResultatService.java
@@ -5,7 +5,7 @@ import org.psychohelp.psychohelp.entity.ResultatTest;
 import java.util.List;
 
 public interface ResultatService {
-    ResultatTest calculerEtEnregistrerResultat(Long citoyenId, Long testId, List<Long> choixIds);
+    ResultatTest calculerEtEnregistrerResultat(Long citoyenId, Integer testId, List<Integer> choixIds);
 
     List<ResultatTest> obtenirResultatsParCitoyen(Long citoyenId);
 }
